Switched to elastic search
Up to this commit the app was using my home made mongoose fulltext
search engine. It was fun to hack it, but compared to ElasticSearch it
was just a toy. So now I use ES.

Subsequently there is no need for mongoose at all, as the app doesn't
store or modify any data. ES as is sufficient as a data store.

Also added check for PORT env var on startup in preparation for Dokku
compatibility.

The single term route is disabled ATM. TODO: enable it back.